+++ src/main/java/com/parkourcraft/parkour/data/SettingsManager.java
@@ -59,8 +59,8 @@ public class SettingsManager {
     public int max_global_personal_completions_leaderboard_size;
     public int max_clans_leaderboard_size;
 
-    public double max_prestige_multiplier;
-    public double prestige_multiplier_per_prestige;
+    public int max_prestige_multiplier;
+    public int prestige_multiplier_per_prestige;
 
     public SettingsManager(FileConfiguration settings) {
         load(settings);
@@ -109,8 +109,8 @@ public class SettingsManager {
         max_global_level_completions_leaderboard_size = settings.getInt("completions.global_level_completions_leaderboard.max_size");
         max_global_personal_completions_leaderboard_size = settings.getInt("completions.global_personal_completions_leaderboard.max_size");
         max_clans_leaderboard_size = settings.getInt("clans.max_leaderboard_size");
-        max_prestige_multiplier = settings.getDouble("prestiges.max_multiplier");
-        prestige_multiplier_per_prestige = settings.getDouble("prestiges.multiplier_per_prestige");
+        max_prestige_multiplier = settings.getInt("prestiges.max_multiplier");
+        prestige_multiplier_per_prestige = settings.getInt("prestiges.multiplier_per_prestige");
         max_rated_levels_leaderboard_size = settings.getInt("levels.max_rated_levels_leaderboard_size");
     }
 }
